Don't redefine the html language
See https://github.com/Microsoft/vscode/issues/26408#issuecomment-301029077: Redefining the html language and giving it a new name causes confusion. A better way is to define a separate language. The django language will be the default language for the .html, but ideally there would be a way to detect of file is html or django-html. 
 
Oh and please enable Issues in your github repo so issues like this can be files.... Thanks a lot!